+++ pubspec.yaml
@@ -30,6 +30,8 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  math_expressions: ^2.3.1
+
 
 
   # The following adds the Cupertino Icons font to your application.
